Refactor out utility function
